+++ src/cpp/r/include/r/RRoutines.hpp
@@ -1,7 +1,7 @@
 /*
  * RRoutines.hpp
  *
- * Copyright (C) 2009-12 by RStudio, Inc.
+ * Copyright (C) 2009-17 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -36,7 +36,7 @@ void registerAll();
       callMethodDef.fun = (DL_FUNC) __NAME__;                                  \
       callMethodDef.numArgs = __NUM_ARGS__;                                    \
       ::rstudio::r::routines::addCallMethod(callMethodDef);                    \
-   } while (0)
+   } while (false)
 
 } // namespace routines   
 } // namespace r