+++ src/gwt/src/org/rstudio/studio/client/server/remote/RemoteServer.java
@@ -36,6 +36,7 @@ import org.rstudio.core.client.jsonrpc.RpcRequest;
 import org.rstudio.core.client.jsonrpc.RpcRequestCallback;
 import org.rstudio.core.client.jsonrpc.RpcResponse;
 import org.rstudio.core.client.jsonrpc.RpcResponseHandler;
+import org.rstudio.studio.client.application.ApplicationTutorialEvent;
 import org.rstudio.studio.client.application.Desktop;
 import org.rstudio.studio.client.application.events.ClientDisconnectedEvent;
 import org.rstudio.studio.client.application.events.EventBus;
@@ -1782,6 +1783,8 @@ public class RemoteServer implements Server
                             String contents,
                             ServerRequestCallback<String> requestCallback)
    {
+      eventBus_.fireEvent(new ApplicationTutorialEvent(ApplicationTutorialEvent.FILE_SAVE));
+
       JSONArray params = new JSONArray();
       params.set(0, new JSONString(id));
       params.set(1, path == null ? JSONNull.getInstance() : new JSONString(path));
@@ -1805,6 +1808,8 @@ public class RemoteServer implements Server
                                 String hash,
                                 ServerRequestCallback<String> requestCallback)
    {
+      eventBus_.fireEvent(new ApplicationTutorialEvent(ApplicationTutorialEvent.FILE_SAVE));
+
       JSONArray params = new JSONArray();
       params.set(0, new JSONString(id));
       params.set(1, path == null ? JSONNull.getInstance() : new JSONString(path));
@@ -1955,6 +1960,7 @@ public class RemoteServer implements Server
                                   String rnwWeave,
                                   ServerRequestCallback<Void> requestCallback)
    {
+      eventBus_.fireEvent(new ApplicationTutorialEvent(ApplicationTutorialEvent.FILE_SAVE));
       JSONArray params = new JSONArray();
       params.set(0, new JSONString(contents));
       params.set(1, JSONBoolean.getInstance(sweave));
@@ -1982,6 +1988,8 @@ public class RemoteServer implements Server
          HashMap<String, String> properties,
          ServerRequestCallback<Void> requestCallback)
    {
+      eventBus_.fireEvent(new ApplicationTutorialEvent(ApplicationTutorialEvent.FILE_SAVE));
+
       JSONObject obj = new JSONObject();
       for (Map.Entry<String, String> entry : properties.entrySet())
       {
@@ -2905,6 +2913,11 @@ public class RemoteServer implements Server
             // first crack goes to globally registered rpc error handlers
             if (!handleRpcErrorInternally(error))
             {
+               eventBus_.fireEvent(new ApplicationTutorialEvent(
+                     ApplicationTutorialEvent.API_ERROR,
+                     "rpc",
+                     error.getEndUserMessage()));
+               
                // no global handlers processed it, send on to caller
                responseHandler.onResponseReceived(RpcResponse.create(error));
             }