+++ NEWS.md
@@ -25,6 +25,10 @@
 * Fix error when quitting while a function named `q()` is present (#1647)
 * Fix crash when executing multiple R Notebook chunks with a failing Rcpp chunk (#1668)
 * Fix missing blank lines in code chunks in R Notebook preview (#1556)
+* Fix selection in Files pane when files are modified while checked (#1715)
+* Fix incorrect truncation of some R object descriptions in Environment pane (#1703)
+* Fix "File Exists" error when using Copy To to overwrite a file (#1722)
+* Fix duplicate prompts in each window when using RStudio API `showPrompt` (#1706)
 
 ### RStudio Server Pro